React testing library and jest
