+++ package.json
@@ -19,6 +19,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^9.1.0",
     "react-scripts": "5.0.1",
+    "react-test-renderer": "^18.2.0",
     "sass": "^1.72.0",
     "tailwind": "^4.0.0",
     "web-vitals": "^2.1.4"
@@ -71,6 +72,7 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^9.0.11",
+    "jest": "^27.5.1",
     "prettier": "^3.1.1"
   }
 }
